Move dataset.py file in order to separate classifiers, tests and settings from modules
