--- rosie/chamber_of_deputies/__main__.py
@@ -1,4 +1,0 @@
-from chamber_of_deputies import main
-
-
-main()
